pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "s2doc"
dynamic = ["version"]
description = "Convert YAML domain models to Markdown documentation"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Igor Music", email = "igor.music@example.com"}
]
keywords = ["domain-driven-design", "ddd", "yaml", "markdown", "documentation", "domain-stories"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Documentation",
    "Topic :: Text Processing :: Markup :: Markdown",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "PyYAML>=6.0",
    "graphviz>=0.20",
    "Pillow>=10.0",
    "lxml>=4.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[project.scripts]
s2doc = "s2doc.cli:main"

[project.urls]
Homepage = "https://github.com/igormusic/domain-stories-visual"
Documentation = "https://github.com/igormusic/domain-stories-visual#readme"
Repository = "https://github.com/igormusic/domain-stories-visual"
Issues = "https://github.com/igormusic/domain-stories-visual/issues"

[tool.setuptools.dynamic]
version = {attr = "s2doc.__version__.__version__"}
